Remove canvas dependencies use mods.
