+++ actions/canvas_draw_image_MOD.ts
@@ -1,5 +1,4 @@
 import type { Action } from '../typings/globals';
-import * as Canvas from 'canvas';
 
 const action: Action<'storage' | 'varName' | 'storage2' | 'varName2' | 'x' | 'y' | 'effect'> = {
   name: 'Canvas: Draw Image on Image',
@@ -67,6 +66,7 @@ const action: Action<'storage' | 'varName' | 'storage2' | 'varName2' | 'x' | 'y'
   },
 
   action(this, cache) {
+    const Canvas = this.getMods().require('canvas');
     const data = cache.actions[cache.index];
     const storage = parseInt(data.storage, 10);
     const varName = this.evalMessage(data.varName, cache);
