+++ actions/canvas_generate_progress_bar_MOD.js
@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-/* eslint-disable @typescript-eslint/no-require-imports */
 module.exports = {
   name: 'Canvas Generate Progress Bar',
   section: 'Image Editing',
@@ -94,7 +92,7 @@ module.exports = {
   },
 
   action(cache) {
-    const Canvas = require('canvas');
+    const Canvas = this.getMods().require('canvas');
     const data = cache.actions[cache.index];
     const storage = parseInt(data.storage, 10);
     const varName = this.evalMessage(data.varName, cache);
