+++ actions/canvas_image_options_MOD.ts
@@ -1,5 +1,4 @@
 import type { Action } from '../typings/globals';
-import * as Canvas from 'canvas';
 
 const action: Action<'storage' | 'varName' | 'mirror' | 'rotation' | 'width' | 'height'> = {
   name: 'Canvas Image Options',
@@ -60,6 +59,7 @@ const action: Action<'storage' | 'varName' | 'mirror' | 'rotation' | 'width' | '
   },
 
   action(this, cache) {
+    const Canvas = this.getMods().require('canvas');
     const data = cache.actions[cache.index];
     const storage = parseInt(data.storage, 10);
     const varName = this.evalMessage(data.varName, cache);
