+++ actions/canvas_send_image_MOD.ts
@@ -1,6 +1,11 @@
-module.exports = {
+import type { Action } from '../typings/globals';
+
+const action: Action<
+  'storage' | 'varName' | 'channel' | 'varName2' | 'message' | 'compress' | 'spoiler' | 'storage2' | 'varName3'
+> = {
   name: 'Canvas Send Image',
   section: 'Image Editing',
+  fields: ['storage', 'varName', 'channel', 'varName2', 'message', 'compress', 'spoiler', 'storage2', 'varName3'],
 
   subtitle(data) {
     const channels = [
@@ -20,8 +25,6 @@ module.exports = {
     if (parseInt(data.storage2, 10) !== varType) return;
     return [data.varName3, 'Message'];
   },
-
-  fields: ['storage', 'varName', 'channel', 'varName2', 'message', 'compress', 'spoiler', 'storage2', 'varName3'],
 
   html(isEvent, data) {
     return `
@@ -91,7 +94,7 @@ module.exports = {
 </div>`;
   },
 
-  init() {
+  init(this: any) {
     const { glob, document } = this;
 
     glob.refreshVariableList(document.getElementById('storage'));
@@ -99,17 +102,15 @@ module.exports = {
     glob.variableChange(document.getElementById('storage2'), 'varNameContainer3');
   },
 
-  action(cache) {
+  action(this, cache) {
     const { DiscordJS } = this.getDBM();
-    const Canvas = require('canvas');
+    const Canvas = this.getMods().require('canvas');
     const data = cache.actions[cache.index];
     const storage = parseInt(data.storage, 10);
     const varName = this.evalMessage(data.varName, cache);
     const imagedata = this.getVariable(storage, varName, cache);
-    if (!imagedata) {
-      this.callNextAction(cache);
-      return;
-    }
+    if (!imagedata) return this.callNextAction(cache);
+
     const channel = parseInt(data.channel, 10);
     const varName2 = this.evalMessage(data.varName2, cache);
     const target = this.getSendTarget(channel, varName2, cache);
@@ -122,8 +123,9 @@ module.exports = {
     const name = `${parseInt(data.spoiler, 10) === 1 ? 'SPOILER_' : ''}image.png`;
     const buffer = canvas.toBuffer('image/png', { compressionLevel: compress });
     const attachment = new DiscordJS.MessageAttachment(buffer, name);
-    if (target && target.send) {
-      target.send(this.evalMessage(data.message, cache), attachment).then((msgobject) => {
+
+    if (target?.send) {
+      target.send(this.evalMessage(data.message, cache), attachment).then((msgobject: any) => {
         const varName3 = this.evalMessage(data.varName3, cache);
         const storage2 = parseInt(data.storage2, 10);
         this.storeValue(msgobject, storage2, varName3, cache);
@@ -136,3 +138,5 @@ module.exports = {
 
   mod() {},
 };
+
+module.exports = action;
