+++ package.json
@@ -16,7 +16,6 @@
   "devDependencies": {
     "@types/jsonpath": "^0.2.0",
     "@types/node": "^16.3.1",
-    "@types/opentype.js": "^1.3.2",
     "@typescript-eslint/eslint-plugin": "^4.28.4",
     "@typescript-eslint/parser": "^4.28.4",
     "discord.js": "^12.5.3",
@@ -35,10 +34,7 @@
     "npm": "7"
   },
   "dependencies": {
-    "canvas": "^2.8.0",
-    "imagedata-filters": "^1.0.1",
     "jsonpath": "^1.1.1",
-    "opentype.js": "^1.3.3",
     "typedoc": "^0.21.4",
     "typedoc-darker-theme": "^1.2.1"
   }
